hf_space/requirements.txt
torch
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.30.0
safetensors
sentencepiece
protobuf
pillow
gradio>=5.0.0
huggingface_hub
spaces
